agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=3.17.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
